+++ anesthetic/read/getdist.py
@@ -2,7 +2,7 @@
 import os
 import re
 import numpy as np
-from anesthetic.samples import MCMCSamples
+from anesthetic.core import MCMCSamples
 from pandas import concat
 
 
@@ -44,7 +44,7 @@ def read_getdist(root, *args, **kwargs):
 
     Returns
     -------
-    :class:`anesthetic.samples.MCMCSamples`
+    :class:`anesthetic.core.MCMCSamples`
 
     """
     dirname, basename = os.path.split(root)
